+++ src/system/spi.h
@@ -83,7 +83,7 @@ void SpiDeInit( Spi_t *obj );
  * \param [IN] cpha  Clock phase
  * \param [IN] slave When set the peripheral acts in slave mode
  */
-void SpiFormat( Spi_t *obj, int8_t bits, int8_t cpol, int8_t cpha, int8_t slave );
+void SpiFormat( Spi_t *obj, uint32_t bits, int8_t cpol, int8_t cpha, int8_t slave );
 
 /*!
  * \brief Sets the SPI speed
